Adding todos to list
The app can now allow users to allow todos to the list. Js creates a <p>, two <button> tags and an <li> element; adds the p tag and buttons to the li element and then adds the li element to the containing Div. After this, the value of the input field is set to an empty string to ready the app for the next task input.