Can delete Todos in list
The application is now able to delete a todo list item.

If the class name of the element clicked inside the #itemList is equal to 'delBtn' then we're clicking on the delete button. In that case, the application deletes every node inside the parent (li element) of this delete button and finally deletes the parent (li element)