Mark box and is box marked functionality
The app can now mark boxes in the table and tell when a box is marked.

When a box isn't marked and is clicked on the class "marked" is added to it. This way it can tell when a box is marked and only increases the "actionCounter" variable every time a non marked box is clicked.

CSS to fix table cell size has also been added (fixed width and height) so cells keep their size regardless if they have content or not.