New function added
THe function gameEndOrMarkedCheck(); was added to be able to duplicate the code for the random "0" mark the programme adds to play against the user.

This is a form of refacgtoring so as to not use duplicate code and reduce the amount of lines in the app and also improve readability and understanding for future modifications.