Added random oponent "O" mark
The "O" mark is randomly added after each "X" mark.

ISSUE: the "O" mark isn't properly detected as a checked item.  This means that you can click and "O" mark and replace it with an "X" mark, and some random "O" marks replace previous "O" marks on the board.

The app doesn't properly detect the "robot" class as a market element, thus overwrites those elements if a new mark is added on that square.