New markerLocation  + limits object
The markerLocation object has been added which includes the current positioin and the start position of the marker.

Also the limits object has been created to access the limits of the board so the marker knows when to stop.

These objects have been added to better deal with the marker position inside functions. It was becoming quite confusing to track the position inside the conditionals.