Bottom limit action fixed
A problem has been fixed. The app wasn't deleting the bottom marker so when running the function getMarkerLocation() the programme was picking up the mark at the bottom of the screen and not continuing the downward movement cell by cell and instead jumping straight to the last cell and reinizializing through resetMarker().