lowerRows() implemented
This function lowers the rows after a line of completed rows has been deleted.

It runs through all the rows above the deleted row to check if there are any bottom-limits, and if it comes across one it lowers it by one row to take the place of the complete row which has previously been deleted.