Access array inside object
An isolated experiment to clarify a problem encountered in the Tetris2 app.

Isolating the problem helped solve the issue at hand.

I was not able to access an array inside an object and change it's element with a double nesting through obj.current.rows[i] and obj.current.columns[i]:

let obj = {

    current: {
        rows: [ ],
        columns: [ ]
    },

     start: {
          row: 0,
          column: 0
     }
}

I had to reduce the nesting of the "current" property and access the nested properties that way through obj.rows[i] and obj.columns[i]:

let obj = {

  rows: [ ],
  columns: [ ],
  start: {
       row: 0,
       column: 0
  }

}