Implementing isMovePossible();
The movement function is being implemented. This function goes determines if the marked items are able to move depending on their current position, other objects surrounding them and the direction they want to move.

For an object to move all its markers must be able to move so the for loop runs through all marked cells and checks if any are unable to move.

If this is the case the movement must abort.