Fixed bug: side movement && bottom-limit
When an object encounters a cell with a class of "bottom-limit" it should not perform the movement. Previous to this bug fix it was performing movement.

The direction conditionals for left and right have been revisited as well as adjustments to the initialization of the variable "nextCell" inside the function isMovePossible();