Implemented new functions
- scanSetObjectContainer(); This function scans the object container after being set to get it's current position values to perform the original position check.

- checkOriginalPosition(); this function checks if the marker is in the original position in relation to the object container. If it isn't the object container doesn't move.

Problem that was solved: when reaching a left limit, the object would move inside the object container and then when pressing right, both the object container and the object would move leaving the object stuck to the far left of the container.

These functions were implemented so that the object container doesn't move until the object is recentered inside the object after toughing a left or right limit.