Bux fix: objectContainer side movement
In the previous version, the object container would paralize movement when it detected the object out of its original position at the same time that it was encountering a left or right limit.

The problem was, when turning the vertical line3 object to a horizontal line3 object, even though the object would be in its original position the object was paralized from the previous vertical line 3 object touching a left or right limit.

The solution was to include an objectContainer.originalPositioin = true; after each turn, because by default the object is returned to it's original position when a new position is generated through a turn.