Bug fixed: lowerRows();
When adding random colours the way cell classes were updated onto the cleared row had to be changed because a random colour has to be taken into account.

Classes can be different from "red", so the solution was to save all the classes for each cell marked "bottom-limit", delete them from the current cell and add them all to the cell directly below it if and only the current row isn't equal to limits.bottom, the lower limit of the board. If that's the case the classes don't get lowered, they just get deleted because there's nowhere to lower them to as the cell following the last cell of the board doesn't exist thus is undefined.