Search functionalities added
1. addSuggestions();

Adds suggestions to the search query the user enters in the input field. If the user clicks away these suggestions dissapear.

2. showFloatingList() + hideFloatingList()

The suggestion list appears when the value of the input field changes. The suggestion list dissapears when the user clicks on one of the list options or when the user clicks an element other than the div element which contains the floating list (it's impossible to click it if you click the list items)

3. fetchDataOnSearch( searchQuery );

has been separated into it's own functino so it takes in a searchQuery argument and sends it to the API. This means the function can be called in buttonSearchHandler() function and also in the searchSuggestionHandler();